+++ python/weka/filters.py
@@ -12,7 +12,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 # filters.py
-# Copyright (C) 2014-2020 Fracpete (pythonwekawrapper at gmail dot com)
+# Copyright (C) 2014-2021 Fracpete (pythonwekawrapper at gmail dot com)
 
 import javabridge
 import logging
@@ -57,10 +57,17 @@ class Filter(OptionHandler):
             jobject = Filter.new_instance(classname)
         self.enforce_type(jobject, "weka.filters.Filter")
         super(Filter, self).__init__(jobject=jobject, options=options)
-        self.__input = javabridge.make_call(self.jobject, "input", "(Lweka/core/Instance;)Z")
-        self.__batchfinished = javabridge.make_call(self.jobject, "batchFinished", "()Z")
-        self.__output = javabridge.make_call(self.jobject, "output", "()Lweka/core/Instance;")
-        self.__outputformat = javabridge.make_call(self.jobject, "getOutputFormat", "()Lweka/core/Instances;")
+
+    def _make_calls(self):
+        """
+        Method for generating instances using javabridge.make_call.
+        Members must start with "_mc_"
+        """
+        super(Filter, self)._make_calls()
+        self._mc_input = javabridge.make_call(self.jobject, "input", "(Lweka/core/Instance;)Z")
+        self._mc_batchfinished = javabridge.make_call(self.jobject, "batchFinished", "()Z")
+        self._mc_output = javabridge.make_call(self.jobject, "output", "()Lweka/core/Instance;")
+        self._mc_outputformat = javabridge.make_call(self.jobject, "getOutputFormat", "()Lweka/core/Instances;")
 
     def capabilities(self):
         """
@@ -89,7 +96,7 @@ class Filter(OptionHandler):
         :return: True if filtered can be collected from output
         :rtype: bool
         """
-        return self.__input(inst.jobject)
+        return self._mc_input(inst.jobject)
 
     def batch_finished(self):
         """
@@ -98,7 +105,7 @@ class Filter(OptionHandler):
         :return: True if instances can be collected from the output
         :rtype: bool
         """
-        return self.__batchfinished()
+        return self._mc_batchfinished()
 
     def outputformat(self):
         """
@@ -107,7 +114,7 @@ class Filter(OptionHandler):
         :return: the output format
         :rtype: Instances
         """
-        inst = self.__outputformat()
+        inst = self._mc_outputformat()
         if inst is None:
             return None
         else:
@@ -120,7 +127,7 @@ class Filter(OptionHandler):
         :return: the filtered instance
         :rtype: an Instance object
         """
-        return Instance(jobject=self.__output())
+        return Instance(jobject=self._mc_output())
 
     def filter(self, data):
         """